Import / export funzionanti

+++ csv2activities/Form1.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using SAPbobsCOM;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -20,21 +21,33 @@ namespace csv2activities
 
         private void button1_Click(object sender, EventArgs e)
         {
+            
             OpenFileDialog dialog = new OpenFileDialog();
             dialog.Filter = "CSV files | *.csv"; // file types, that will be allowed to upload
             dialog.Filter += "| TXT files | *.txt"; 
             dialog.Multiselect = false; // allow/deny user to upload more than one file at a time
             if (dialog.ShowDialog() == DialogResult.OK) // if user clicked OK
             {
+                var myCompany = connectSap();
+
                 string path = dialog.FileName; // get name of file
-                funzioniComuni.fetchCSV(path,';');
+                bool check = SAPHelper.insActFromCSV(myCompany,path,',');
+
+                if (check)
+                {
+                    MessageBox.Show("All records are imported in SAP B1", "Import from CSV", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                else {
+                    MessageBox.Show("Somethings goes wrong, check the CSV format", "Import from CSV", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+
+                SAPHelper.destroyCom(myCompany);
 
             }
-
-            
+         
         }
 
-        public void connectSap() {
+        public Company connectSap(bool destroy=false, bool popup = false) {
 
             string server = textBox1.Text;
             string port = textBox2.Text;
@@ -69,25 +82,42 @@ namespace csv2activities
             }
             
             //connetti a DB
-            
-            //var myCompany = SAPHelper.getSocieta("hana", "30015", tipo, "UTENTESAPB1", "Passdb01", "manager", "manager", "SBODEMOIT");
-            var myCompany = SAPHelper.getSocieta(server, port, tipo, dbUser, dbPassword, b1User, b1Pass, b1Company);
-            if (myCompany != null)
+           
+            Company myCompany = SAPHelper.getSocieta(server, port, tipo, dbUser, dbPassword, b1User, b1Pass, b1Company);
+            if (popup)
             {
-                MessageBox.Show("Correct configuration", "Connection Test", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                if (myCompany != null)
+                {
+                    MessageBox.Show("Correct configuration", "Connection Test", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                else
+                {
+                    MessageBox.Show("Some data is missing or wrong", "Connection Test", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+
             }
-            else {
-                MessageBox.Show("Some data is missing or wrong", "Connection Test", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            }
+
 
             //SAPHelper.getAct(myCompany);
-            SAPHelper.destroyCom(myCompany);
+            if (destroy)
+            {
+                SAPHelper.destroyCom(myCompany);
+                return null;
+            }
 
+            return myCompany;
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
-            connectSap();
+            connectSap(true, true);
+        }
+
+        private void button2_Click(object sender, EventArgs e)
+        {
+            var myCompany = connectSap();
+            SAPHelper.getActToXML(myCompany);
+            SAPHelper.destroyCom(myCompany);
         }
     }
 }
